requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cloudinary-storage==0.3.0
ipython==8.19.0
python-dotenv==1.0.0
Pillow==10.2.0
dj-database-url==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9

